format:将changelog抽离成组件
